+++ scripts/process_level_1.py
@@ -3,13 +3,11 @@ import os
 import yaml
 import xarray as xr
 import shutil
-from src.process_lev0 import (
+from src.process import (
     radiometer,
     radar,
     iwv,
     add_georeference,
-)
-from src.process_lev1 import (
     correct_radar_height,
     filter_radar,
     add_masks_radar,
